Refactor the Participant into a separate class.
